hopefuly fixed cmmmotionmanager

+++ initialziji/ViewController.h
@@ -23,7 +23,7 @@
 
 //@property (nonatomic) NSInteger YawDegrees;
 
-+(double)readIt;
++(float)readIt;
